+++ src/nav_solution.cpp
@@ -1,6 +1,6 @@
 #include "nav_solution.h"
 
-const float& Nav_solution::vel(int num) { return v(num); }
+const float& Nav_solution::vel(size_t num) { return v(num); }
 
 const matrix::Vector3f& Nav_solution::vel() { return v; }
 
@@ -10,7 +10,7 @@ void Nav_solution::vel(float vec_in[3]) {
   }
 }
 
-const float& Nav_solution::rot(int num) { return r(num); }
+const float& Nav_solution::rot(size_t num) { return r(num); }
 
 const matrix::Eulerf& Nav_solution::rot() { return r; }
 
@@ -20,7 +20,7 @@ void Nav_solution::rot(float vec_in[3]) {
   }
 }
 
-const float& Nav_solution::pos(int num) { return p(num); }
+const float& Nav_solution::pos(size_t num) { return p(num); }
 
 const matrix::Vector2f& Nav_solution::pos() { return p; }
